fix: extract workspace clean logic to _workspace.py for line limit
cli/app.py exceeded the 500-line architectural guard (CD1) after the
age-based safety guards were added. Extract _format_age and
run_workspace_clean into cli/_workspace.py, keeping the cyclopts
command handler as a thin delegate in app.py.
